+++ alpha/apps/kaltura/lib/request/kSessionBase.class.php
@@ -84,6 +84,8 @@ class kSessionBase
 	const RANDOM_SIZE = 16;
 	const AES_IV = "\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0\0";	// no need for an IV since we add a random string to the message anyway
 	
+	const FIVE_MINUTES_IN_SECONDS = 300;
+	
 	const FIELD_EXPIRY =              '_e';
 	const FIELD_TYPE =                '_t';
 	const FIELD_USER =                '_u';
@@ -467,8 +469,16 @@ class kSessionBase
 		$rand = null;
 		$expiry += time();
 		
+		if(strpos($privileges, kSessionBase::PRIVILEGE_DOWNLOAD_ASSET) !== false)
+		{
+			$expiry = ceil($expiry/self::FIVE_MINUTES_IN_SECONDS) * self::FIVE_MINUTES_IN_SECONDS;
+			$rand = $expiry;
+		}
+		
 		if ($ksVersion == 2)
+		{
 			return self::generateKsV2($adminSecretForSigning, $userId, $type, $partnerId, $expiry, $privileges, $masterPartnerId, $additionalData, $rand);
+		}
 
 		return self::generateKsV1($adminSecretForSigning, $userId, $type, $partnerId, $expiry, $privileges, $masterPartnerId, $additionalData, $rand);
 	}