+++ plugins/content/entry_permission_level/lib/api/KalturaPermissionLevelUserEntry.php
@@ -1,0 +1,55 @@
+<?php
+/**
+ * @package plugins.entryPermissionLevel
+ * @subpackage api
+ */
+class KalturaPermissionLevelUserEntry extends KalturaUserEntry
+{
+	/**
+	 * Playback context
+	 * @var KalturaPermissionLevelArray
+	 */
+	public $permissionLevels;
+	
+	/**
+	 * mapping between the field on this object (on the left) and the setter/getter on the entry object (on the right)
+	 */
+	private static $map_between_objects = array('permissionLevels');
+	
+	/* (non-PHPdoc)
+	 * @see KalturaObject::getMapBetweenObjects()
+	 */
+	public function getMapBetweenObjects()
+	{
+		return array_merge(parent::getMapBetweenObjects(), self::$map_between_objects);
+	}
+	
+	/* (non-PHPdoc)
+	 * @see KalturaObject::toObject()
+	 */
+	public function toObject($dbObject = null, $propertiesToSkip = array())
+	{
+		if(is_null($dbObject))
+			$dbObject = new PermissionLevelUserEntry();
+		
+		return parent::toObject($dbObject, $propertiesToSkip);
+	}
+	
+	/* (non-PHPdoc)
+	 * @see KalturaObject::toInsertableObject()
+	 */
+	public function toInsertableObject ( $object_to_fill = null , $props_to_skip = array() )
+	{
+		$object_to_fill = parent::toInsertableObject($object_to_fill, $props_to_skip);
+		if (kCurrentContext::getCurrentSessionType() == SessionType::USER)
+		{
+			if ($this->userId && (!kCurrentContext::getCurrentKsKuser() ||strtolower(kCurrentContext::getCurrentKsKuser()->getPuserId()) != strtolower($this->userId)))
+			{
+				throw new KalturaAPIException (KalturaErrors::INVALID_USER_ID);
+			}
+		}
+		
+		return $object_to_fill;
+	}
+	
+}