+++ plugins/content/entry_permission_level/lib/api/enums/KalturaUserEntryPermissionLevel.php
@@ -1,0 +1,9 @@
+<?php
+/**
+ * @package plugins.entryPermissionLevel
+ * @subpackage api.enum
+ */
+class KalturaUserEntryPermissionLevel extends KalturaEnum implements UserEntryPermissionLevel
+{
+
+}