+++ plugins/content/entry_permission_level/lib/api/filters/KalturaPermissionLevelUserEntryFilter.php
@@ -1,0 +1,14 @@
+<?php
+/**
+ * @package plugins.entryPermissionLevel
+ * @subpackage api.filters
+ */
+class KalturaPermissionLevelUserEntryFilter extends KalturaUserEntryFilter
+{
+	public function getListResponse(KalturaFilterPager $pager, KalturaDetachedResponseProfile $responseProfile = null)
+	{
+		$this->typeEqual = RegistrationPlugin::getApiValue(PermissionLevelUserEntryType::PERMISSION_LEVEL);
+		$response = parent::getListResponse($pager, $responseProfile);
+		return $response;
+	}
+}