+++ plugins/content/entry_permission_level/lib/model/PermissionLevel.php
@@ -1,0 +1,19 @@
+<?php
+/**
+ * @package plugins.entryPermissionLevel
+ * @subpackage model
+ */
+class PermissionLevel extends UserEntry
+{
+	protected $permissionLevel;
+	
+	public function getPermissionLevel()
+	{
+		return $this->permissionLevel;
+	}
+	
+	public function setPermissionLevel($permissionLevel)
+	{
+		$this->permissionLevel = $permissionLevel;
+	}
+}