+++ release-notes.md
@@ -1,3 +1,24 @@
+# Rigel-18.12.0
+
+## Support userEntry object to define permission level of user in specific entry context ##
+- Issue Type: Feature
+- Issue ID: 
+
+Enable plugin:
+
+	To enable this feature plugin add the following to your plugins.ini file:
+	- EntryPermissionLevel
+
+### Deployment Scripts ###
+    php /opt/kaltura/app/deployment/base/scripts/installPlugins.php
+
+## Add systemPartner getConfiguration permission for self serve partner ##
+* Issue Type: Task
+* Issue ID: SSRV-710
+
+### Deployment Scripts ###
+    php deployment/updates/scripts/add_permissions/2022_08_09_self_serve_add_permission_systempartner_getconfiguration.php
+
 # Rigel-18.11.0
 
 ## Add partner for BI ##